+++ tests/pna/layout/test_layout.py
@@ -49,13 +49,10 @@ def test_layout(pna_pxl_dataset: PNAPixelDataset, tmp_path):
 
 @pytest.mark.slow
 def test_layout_from_path(pna_pxl_dataset: PNAPixelDataset, tmp_path):
-    manager = AnalysisManager(
-        [
-            CreateLayout(
-                ["wpmds_3d"],
-            )
-        ]
+    layout_task = CreateLayout(
+        ["wpmds_3d"],
     )
+    manager = AnalysisManager([layout_task])
 
     pna_pixel_filtered = pna_pxl_dataset.filter(
         components=pna_pxl_dataset.adata().obs.index[:2]
@@ -83,9 +80,5 @@ def test_layout_from_path(pna_pxl_dataset: PNAPixelDataset, tmp_path):
     assert "A" in set(result["pixel_type"].to_list())
     assert "B" in set(result["pixel_type"].to_list())
 
-    assert len(manager._temp_folders_used) == 1, (
-        "Expected one temporary folder to be used."
-    )
-    assert not os.path.exists(manager._temp_folders_used[0]), (
-        f"Temporary folder {manager._temp_folders_used[0]} was not cleaned up."
-    )
+    # Make sure we have clean out the work directory
+    assert not layout_task.get_work_folder().exists()  # type: ignore